complete sync blocks
what a journey. add test for a late comer in server_test.

one concerning hangover from this is that response addresses are leaking
into the messages. i can't tell if this is side effect of the
localTransport layer or if it's something else. perhaps it's simply
that i'm not usually in the transport layer, so i'm not used to explicitly
tracking how to route messages