Add tcp transport test
this was labrious because we had been tracking From net.Addr in the RPC

this doesnt work well for a number of reasons. the biggest is serialization
of a net.Addr across the network. ultimately, we didn't need net.Addr, we
were only using it as a key to map - a string id works fine.

other than that, had to implement message parsing on top of the tcp protocal

there is much room for improvement of factor out the current parsing from the read loop into some testable funcs