make block syncing work with tcp transport
the crux here, again, was  bad interface choice in the transport layer
at the beginning of the project.

the server needs to send blocks to somewhere. in tcp, the connections for this
communication are hidden in the transport layer. this is fine and good. the
problem is that since they are hidden, we need bookkeeping to map from
some peer-server identifier to the actual connection to be use for sending

the original transport layer intermixed the notion of connections and transportation and made implicit assumptions about the mapping from peer-server id to connection. it got away with this because the local transport did not have any notion of connection -- the transport was the connection and it was 1:1. in tcp we have a listener and spawned connections. the transport is not 1:1 with a connection.